+++ src/shared/lib/hooks/http-hook/http.hook.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState, useCallback } from 'react';
 
 type HttpMethod = 'GET' | 'POST' | 'PUT' | 'DELETE';
