+++ src/shared/lib/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './use-click-outside/use-click-outside';
 export * from './use-app-state/use-app-state';
+export * from './http-hook/http.hook';
