+++ src/shared/services/index.ts
@@ -1,0 +1,1 @@
+export * from './get-games/get-games';
